+++ packages/api-client/src/errors.ts
@@ -45,7 +45,7 @@ export function isRecoverable(error) {
   return true;
 }
 
-export function getVersionIncompatibleMessage(error: ServerResponseError, response: Response) {
+export function getVersionIncompatibleMessage(error: { message: string }, response: Response) {
   if (error.message === VERSION_COMPATIBILITY_ERRORS.LOW) {
     return 'Tamanu is out of date, reload to get the new version! If that does not work, contact your system administrator.';
   }
