+++ packages/api-client/src/types.ts
@@ -5,19 +5,3 @@ interface Logger {
   }
   
   export type LoggerType = Logger | Console;
-
-
-export interface FetchOptions extends RequestInit {
-    fetch?: typeof fetch;
-    timeout?: number | false;
-  }
-
-  export interface ResponseError {
-    name: string;
-    message: string;
-  }
-  
-  export interface ResponseErrorData {
-    error?: ResponseError;
-    [key: string]: any;
-  }
